+++ app/controllers/static_pages_controller.rb
@@ -1,11 +1,5 @@
 class StaticPagesController < ApplicationController
   def home
-    # @micropost = current_user.microposts.build if logged_in?
-
-    if logged_in?
-      @micropost  = current_user.microposts.build
-      @feed_items = current_user.feed.paginate(page: params[:page])
-    end
   end
 
   def help
