+++ app/controllers/users_controller.rb
@@ -9,7 +9,6 @@ class UsersController < ApplicationController
   def show
     @user = User.find(params[:id])
     # debugger
-    @microposts = @user.microposts.paginate(page: params[:page])
   end
   def new
     @user = User.new
@@ -64,13 +63,13 @@ class UsersController < ApplicationController
   # beforeフィルタ
 
   # ログイン済みユーザーかどうか確認
-  # def logged_in_user
-  #   unless logged_in?
-  #     store_location
-  #     flash[:danger] = "Please log in."
-  #     redirect_to login_url, status: :see_other
-  #   end
-  # end
+  def logged_in_user
+    unless logged_in?
+      store_location
+      flash[:danger] = "Please log in."
+      redirect_to login_url, status: :see_other
+    end
+  end
 
   # 正しいユーザーかどうか確認
   def correct_user
