+++ app/models/user.rb
@@ -1,6 +1,4 @@
 class User < ApplicationRecord
-  has_many :microposts, dependent: :destroy
-
   # attr_accessor :remember_token, :activation_token
   attr_accessor :remember_token, :activation_token, :reset_token
   before_save   :downcase_email
@@ -85,12 +83,6 @@ class User < ApplicationRecord
   # パスワード再設定の期限が切れている場合はtrueを返す
   def password_reset_expired?
     reset_sent_at < 2.hours.ago
-  end
-
-  # 試作feedの定義
-  # 完全な実装は次章の「ユーザーをフォローする」を参照
-  def feed
-    Micropost.where("user_id = ?", id)
   end
 
   private
