+++ app/views/password_resets/edit.html.erb
@@ -4,7 +4,7 @@
 <div class="row">
   <div class="col-md-6 col-md-offset-3">
     <%= form_with(model: @user, url: password_reset_path(params[:id])) do |f| %>
-      <%= render 'shared/error_messages', object: f.object %>
+      <%= render 'shared/error_messages' %>
 
       <%= hidden_field_tag :email, @user.email %>
