+++ app/views/shared/_error_messages.html.erb
@@ -1,10 +1,10 @@
-<% if object.errors.any? %>
+<% if @user.errors.any? %>
   <div id="error_explanation">
     <div class="alert alert-danger">
-      The form contains <%= pluralize(object.errors.count, "error") %>.
+      The form contains <%= pluralize(@user.errors.count, "error") %>.
     </div>
     <ul>
-      <% object.errors.full_messages.each do |msg| %>
+      <% @user.errors.full_messages.each do |msg| %>
         <li><%= msg %></li>
       <% end %>
     </ul>
