--- app/views/shared/_feed.html.erb
@@ -1,7 +1,0 @@
-<% if @feed_items.any? %>
-  <ol class="microposts">
-    <%= render @feed_items %>
-  </ol>
-  <%= will_paginate @feed_items,
-                    params: { controller: :static_pages, action: :home } %>
-<% end %>
