+++ app/views/users/new.html.erb
@@ -4,7 +4,7 @@
 <div class="row">
   <div class="col-md-6 col-md-offset-3">
     <%= form_with(model: @user) do |f| %>
-      <%= render 'shared/error_messages', object: f.object %>
+      <%= render 'shared/error_messages' %>
 
       <%= f.label :name %>
       <%= f.text_field :name, class: 'form-control' %>
