+++ config/routes.rb
@@ -19,6 +19,4 @@ Rails.application.routes.draw do
   resources :users
   resources :account_activations, only: [:edit]
   resources :password_resets,     only: [:new, :create, :edit, :update]
-  resources :microposts,          only: [:create, :destroy]
-  get '/microposts', to: 'static_pages#home'
 end
