+++ db/schema.rb
@@ -10,16 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_05_30_180535) do
-  create_table "microposts", force: :cascade do |t|
-    t.text "content"
-    t.integer "user_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.index ["user_id", "created_at"], name: "index_microposts_on_user_id_and_created_at"
-    t.index ["user_id"], name: "index_microposts_on_user_id"
-  end
-
+ActiveRecord::Schema[7.0].define(version: 2023_05_30_173311) do
   create_table "users", force: :cascade do |t|
     t.string "name"
     t.string "email"
@@ -36,5 +27,4 @@ ActiveRecord::Schema[7.0].define(version: 2023_05_30_180535) do
     t.index ["email"], name: "index_users_on_email", unique: true
   end
 
-  add_foreign_key "microposts", "users"
 end
