+++ db/seeds.rb
@@ -19,10 +19,3 @@ User.create!(name:  "Example User",
                activated: true,
                activated_at: Time.zone.now)
 end
-
-# ユーザーの一部を対象にマイクロポストを生成する
-users = User.order(:created_at).take(6)
-50.times do
-  content = Faker::Lorem.sentence(word_count: 5)
-  users.each { |user| user.microposts.create!(content: content) }
-end
