--- test/fixtures/microposts.yml
@@ -1,41 +1,0 @@
-orange:
-  content: "I just ate an orange!"
-  created_at: <%= 10.minutes.ago %>
-  user: michael
-
-tau_manifesto:
-  content: "Check out the @tauday site by @mhartl: https://tauday.com"
-  created_at: <%= 3.years.ago %>
-  user: michael
-
-cat_video:
-  content: "Sad cats are sad: https://youtu.be/PKffm2uI4dk"
-  created_at: <%= 2.hours.ago %>
-  user: michael
-
-most_recent:
-  content: "Writing a short test"
-  created_at: <%= Time.zone.now %>
-  user: michael
-
-ants:
-  content: "Oh, is that what you want? Because that's how you get ants!"
-  created_at: <%= 2.years.ago %>
-  user: archer
-
-zone:
-  content: "Danger zone!"
-  created_at: <%= 3.days.ago %>
-  user: archer
-
-tone:
-  content: "I'm sorry. Your words made sense, but your sarcastic tone did not."
-  created_at: <%= 10.minutes.ago %>
-  user: lana
-
-<% 30.times do |n| %>
-micropost_<%= n %>:
-  content: <%= Faker::Lorem.sentence(word_count: 5) %>
-  created_at: <%= 42.days.ago %>
-  user: michael
-<% end %>
